[Nexmark] Extract BidGenerator from Generator
