+++ sdks/java/nexmark/src/main/java/org/apache/beam/sdk/nexmark/sources/generator/model/AuctionGenerator.java
@@ -16,19 +16,18 @@
  * limitations under the License.
  */
 
-package org.apache.beam.sdk.nexmark.sources.utils;
+package org.apache.beam.sdk.nexmark.sources.generator.model;
 
-import static org.apache.beam.sdk.nexmark.sources.utils.LongGenerator.nextLong;
-import static org.apache.beam.sdk.nexmark.sources.utils.PersonGenerator.lastBase0PersonId;
-import static org.apache.beam.sdk.nexmark.sources.utils.PersonGenerator.nextBase0PersonId;
-import static org.apache.beam.sdk.nexmark.sources.utils.PriceGenerator.nextPrice;
-import static org.apache.beam.sdk.nexmark.sources.utils.StringsGenerator.nextExtra;
-import static org.apache.beam.sdk.nexmark.sources.utils.StringsGenerator.nextString;
+import static org.apache.beam.sdk.nexmark.sources.generator.model.LongGenerator.nextLong;
+import static org.apache.beam.sdk.nexmark.sources.generator.model.PersonGenerator.lastBase0PersonId;
+import static org.apache.beam.sdk.nexmark.sources.generator.model.PersonGenerator.nextBase0PersonId;
+import static org.apache.beam.sdk.nexmark.sources.generator.model.PriceGenerator.nextPrice;
+import static org.apache.beam.sdk.nexmark.sources.generator.model.StringsGenerator.nextExtra;
+import static org.apache.beam.sdk.nexmark.sources.generator.model.StringsGenerator.nextString;
 
 import java.util.Random;
-
 import org.apache.beam.sdk.nexmark.model.Auction;
-import org.apache.beam.sdk.nexmark.sources.GeneratorConfig;
+import org.apache.beam.sdk.nexmark.sources.generator.GeneratorConfig;
 
 /**
  * AuctionGenerator.
@@ -76,7 +75,7 @@ public class AuctionGenerator {
     String desc = nextString(random, 100);
     long reserve = initialBid + nextPrice(random);
     int currentSize = 8 + name.length() + desc.length() + 8 + 8 + 8 + 8 + 8;
-    String extra = nextExtra(random, currentSize, config.configuration.avgAuctionByteSize);
+    String extra = nextExtra(random, currentSize, config.getAvgAuctionByteSize());
     return new Auction(id, name, desc, initialBid, reserve, timestamp, expires, seller, category,
         extra);
   }
@@ -128,18 +127,16 @@ public class AuctionGenerator {
     long currentEventNumber = config.nextAdjustedEventNumber(eventsCountSoFar);
     // How many events till we've generated numInFlightAuctions?
     long numEventsForAuctions =
-        (config.configuration.numInFlightAuctions * GeneratorConfig.PROPORTION_DENOMINATOR)
+        (config.getNumInFlightAuctions() * GeneratorConfig.PROPORTION_DENOMINATOR)
             / GeneratorConfig.AUCTION_PROPORTION;
     // When will the auction numInFlightAuctions beyond now be generated?
-    long futureAuction =
-        config.timestampAndInterEventDelayUsForEvent(currentEventNumber + numEventsForAuctions)
-            .getKey();
+    long futureAuction = config
+        .timestampAndInterEventDelayUsForEvent(currentEventNumber + numEventsForAuctions)
+        .getKey();
     // System.out.printf("*** auction will be for %dms (%d events ahead) ***\n",
     //     futureAuction - timestamp, numEventsForAuctions);
     // Choose a length with average horizonMs.
     long horizonMs = futureAuction - timestamp;
     return 1L + nextLong(random, Math.max(horizonMs * 2, 1L));
   }
-
-
 }
