+++ sdks/java/nexmark/src/main/java/org/apache/beam/sdk/nexmark/sources/generator/model/package-info.java
@@ -17,6 +17,6 @@
  */
 
 /**
- * Utility classes for Generator.
+ * Model Generators.
  */
-package org.apache.beam.sdk.nexmark.sources.utils;
+package org.apache.beam.sdk.nexmark.sources.generator.model;
