+++ sdks/java/nexmark/src/main/java/org/apache/beam/sdk/nexmark/sources/generator/package-info.java
@@ -16,17 +16,11 @@
  * limitations under the License.
  */
 
-package org.apache.beam.sdk.nexmark.sources.utils;
-
-import java.util.Random;
+/**
+ * org.apache.beam.sdk.nexmark.sources.generator.
+ */
 
 /**
- * Generates a random price.
+ * Events generation logic.
  */
-public class PriceGenerator {
-
-  /** Return a random price. */
-  public static long nextPrice(Random random) {
-    return Math.round(Math.pow(10.0, random.nextDouble() * 6.0) * 100.0);
-  }
-}
+package org.apache.beam.sdk.nexmark.sources.generator;
